requirements.txt
django==4.2.*
python-telegram-bot==20.6.*
aiohttp==3.8.*
environs==9.5.*
django-relativedelta==2.0.*
dj-database-url==2.1.0
requests==2.31.*
celery==5.3.*
redis==5.0.*
pandas==2.2.2
playwright==1.43.0
pre-commit==3.7.0
django-celery-beat==2.6.0
psycopg2-binary==2.9.9
flower==2.0.1
TikTokApi==6.3.0
more-itertools==10.2.0
